fix race condition when listening to channels
In the past if a consumer sent an error while RunPipeline was sending
data to a consumer channel, the pipeline would deadlock. RunPipeline
would be waiting on consumer to accept data, while consumer was waiting
on RunPipeline to accept an error.

Split the listeners out of a select into their own goroutines, each able
to send an error to a loop that blocks until an error is received and
returned, or until the error channel is closed when we have closed errs