add support for remote producers
remote producers describe a producer who yields producers. The idea is
that a pipeline can load a source up with data, and for every piece of data,
we extrapolate a discreet producer to source it's own pipeline

on pipelines a `produce-from` field can be supplied with a producer. For now,
it only fetches the first item from the producer