No hcl body decoding here
Since we get the mover options as cty values that are already decoded,
the hcl body decoding was pretty redundant. In the future, we'll no
longer make a parser, we'll just decode with regular `cty:` tags

Building pipelines no longer need an evalctx. The one responsible for
meta-producers needs the library context