Use a channel to spawn producers
When compiling pipelines into something runnable, instead of getting
producers ( remote and literal ) into a function that's immediately
callable, use a function that returns a channel that will spawn some
producers before closing.

With this channel we can feed literal producers as normal, while
allowing us to get remote producer descriptors during runtime, without
any timeout.

Refactor RunPipeline to collect producers from the spawner channel in
chunks, running them in the order received, until the spawner closes (
which we test for by looking at whether or not the next producer is nil
)

Refactor the pipeline testcase struct to specify how many producers to
run in parallel so that we can test for that, testing for non-wholly
divisible and overflowing parallel counts